backend/requirements.txt
asgiref==3.7.2
Django==5.0.3
django-model-utils==4.4.0
sqlparse==0.4.4
